ci: add Depends-on hint for dependent PRs in PR description
There can be scenarios where a PR from one kernel topic can rely on a PR
from another kernel topic branch. To enable proper builds in the CI,
this commit introduces a "Depends-on" hint in the pull request
description.

When a CI job is run, the pull request description will be parsed for a
"Depends-on" hint to pick up the list of dependent PRs. The dependent
PRs are filtered out if they haven't yet been merged in their respective
topic branches.

Eg. Hint "Depends-on: #22, #44"

Signed-off-by: Yashasvi Nancherla <ynancher@qti.qualcomm.com>
Signed-off-by: Abin Joseph <abinjose@qti.qualcomm.com>