Move coral template from .github/workflows to .github/templates
This pull request separates the coral template to a separate templates directory
instead of having in the workflows directory. The issue with the template having
 in the workflows directory is that the github treats every .yml file inside
the .github/workflows directory as a workflow file, having the coral template
in the same directory was a issue for the github actions UI. This eliminates
the unnecessary workflow lists.

Signed-off-by: Vishal Kumar <viskuma@qti.qualcomm.com>