Update base_sha and head_sha for kernel checkers
This commit ensures that the `head_sha` and the `base_sha` are not passed as input
to the checker action, instead the head and base shas' are determined
after the workspace is ready, which ensures that proper changes history
is passed to the checker script for analysis.

This commit removes the `prepare` job from the workflow
it isnn't not required now. The similar commit to `kernel-checkers` repo is
also made to accommodate changes done in using the checker workflow.

Signed-off-by: Hasibur Rahman Mohammed <hasiburr@qti.qualcomm.com>