Update final status in pre-merge as per build for kernel-topics
This commit updates the workflow status based on conditions.
RCA:
The workflow status on the PR used to update the status as
failed for topics repo as the tests are disabled for the topics.

Fix:
Tests are skipped for kernel-topics
adding check to prevent workflow failure
if tests get skipped for topics repo.

Signed-off-by: ynancher <ynancher@qti.qualcomm.com>